Assign a type to all AST nodes.
We're venturing into uncharted territory here, so I'm not totally sure this is
correct.

We assign a term to a type when we infer a type for it, and we apply new
solutions from the output context after checking a term. One wrinkle is that
when we check the type of a lambda we don't first infer a type for it (like we
do when checking all other terms), so in that case we first assign the
to-be-checked-against type as the lambda term's type, and then refine it with
solutions as usual when checking is complete.