Added if expressions.
I'm not completely sure the typing approach is correct. When synthesizing a
type for an if expression, I synthesize a type for the true branch and then
check the false branch against it, and return the output context from that
check. There may be extra information in that context that needs to be peeled
off, which I'm not doing.

I also explicitly check if AST nodes rather than using the default "infer and
then subsume" case, because it seems that if we have a type to check against we
should just check both branches against that type directly, rather than
synthesize, check and then subsume.